Corrections in euler_flat_earth style & fixed test not passing in python 2

+++ src/pyfme/models/tests/test_euler_flat_earth.py
@@ -17,15 +17,15 @@ from pyfme.models.euler_flat_earth import (linear_and_angular_momentum_eqs,
 def test1_linear_and_angular_momentum_eqs():
 
     time = 0
-    vel = np.array([1, 1, 1, 1, 1, 1])
+    vel = np.array([1, 1, 1, 1, 1, 1], dtype=float)
     mass = 10
     inertia = np.array([[1000,    0, -100],
                         [   0,  100,    0],
-                        [-100,    0,  100]])
-    forces = np.array([100, 100, 100])
-    moments = np.array([100, 1000, 100])
+                        [-100,    0,  100]], dtype=float)
+    forces = np.array([100., 100., 100.], dtype=float)
+    moments = np.array([100., 1000., 100], dtype=float)
 
-    expected_sol = np.array([10, 10, 10, 11/9, 1, 92/9])
+    expected_sol = np.array([10, 10, 10, 11./9, 1, 92./9], dtype=float)
     sol = linear_and_angular_momentum_eqs(time, vel, mass, inertia, forces,
                                           moments)
     assert(np.allclose(expected_sol, sol))
@@ -34,15 +34,15 @@ def test1_linear_and_angular_momentum_eqs():
 def test2_linear_and_angular_momentum_eqs():
 
     time = 0
-    vel = np.array([1, 0, 1, 0, 1, 0])
+    vel = np.array([1, 0, 1, 0, 1, 0], dtype=float)
     mass = 10
     inertia = np.array([[100,    0, -10],
                         [  0,  100,   0],
-                        [-10,    0, 100]])
-    forces = np.array([1000, 10, 10])
-    moments = np.array([100, 100, 100])
+                        [-10,    0, 100]], dtype=float)
+    forces = np.array([1000, 10, 10], dtype=float)
+    moments = np.array([100, 100, 100], dtype=float)
 
-    expected_sol = np.array([99, 1, 2, 10./9, 1, 10./9])
+    expected_sol = np.array([99, 1, 2, 10./9, 1, 10./9], dtype=float)
     sol = linear_and_angular_momentum_eqs(time, vel, mass, inertia, forces,
                                           moments)
     assert(np.allclose(expected_sol, sol))
@@ -51,13 +51,13 @@ def test2_linear_and_angular_momentum_eqs():
 def test1_jac_linear_and_angular_momentum_eqs():
 
     time = 0
-    vel = np.array([1, 1, 1, 1, 1, 1])
+    vel = np.array([1, 1, 1, 1, 1, 1], dtype=float)
     mass = 10
     inertia = np.array([[1000,    0, -100],
                         [   0,  100,    0],
-                        [-100,    0,  100]])
+                        [-100,    0,  100]], dtype=float)
 
-    expected_sol = np.zeros([6, 6])
+    expected_sol = np.zeros([6, 6], dtype=float)
 
     expected_sol[0, 1] = 1
     expected_sol[0, 2] = - 1
@@ -93,13 +93,13 @@ def test1_jac_linear_and_angular_momentum_eqs():
 def test2_jac_linear_and_angular_momentum_eqs():
 
     time = 0
-    vel = np.array([1, 0, 1, 0, 1, 0])
+    vel = np.array([1, 0, 1, 0, 1, 0], dtype=float)
     mass = 10
     inertia = np.array([[100,    0, -10],
                         [  0,  100,   0],
-                        [-10,    0, 100]])
+                        [-10,    0, 100]], dtype=float)
 
-    expected_sol = np.zeros([6, 6])
+    expected_sol = np.zeros([6, 6], dtype=float)
 
     expected_sol[0, 2] = - 1
     expected_sol[0, 4] = - 1
@@ -111,7 +111,7 @@ def test2_jac_linear_and_angular_momentum_eqs():
     expected_sol[2, 4] = 1
 
     expected_sol[3, 3] = 10./99
-    expected_sol[3, 5] = - 1/99
+    expected_sol[3, 5] = - 1./99
 
     expected_sol[5, 3] = 1./99
     expected_sol[5, 5] = - 10./99
@@ -125,7 +125,7 @@ def test1_kinematic_angular_eqs():
 
     time = 0
     euler_angles = np.array([np.pi / 4, np.pi / 4, 0])
-    angular_vel = np.array([1, 1, 1])
+    angular_vel = np.array([1, 1, 1], dtype=float)
 
     expected_sol = np.array([0, 1 + 2 ** 0.5, 2])
     sol = kinematic_angular_eqs(time, euler_angles, angular_vel)
@@ -137,9 +137,9 @@ def test2_kinematic_angular_eqs():
 
     time = 0
     euler_angles = np.array([0, np.pi / 2, 0])
-    angular_vel = np.array([0, 1, 0])
+    angular_vel = np.array([0, 1, 0], dtype=float)
 
-    expected_sol = np.array([0, 0, 1])
+    expected_sol = np.array([0, 0, 1], dtype=float)
     sol = kinematic_angular_eqs(time, euler_angles, angular_vel)
 
     assert(np.allclose(expected_sol, sol))
@@ -149,7 +149,7 @@ def test1_jac_kinematic_angular_eqs():
 
     time = 0
     euler_angles = np.array([np.pi / 4, np.pi / 4, 0])
-    angular_vel = np.array([1, 1, 1])
+    angular_vel = np.array([1, 1, 1], dtype=float)
 
     expected_sol = np.zeros([3, 3])
     expected_sol[0, 1] = - 2 ** 0.5
@@ -165,9 +165,9 @@ def test2_jac_kinematic_angular_eqs():
 
     time = 0
     euler_angles = np.array([0, np.pi / 2, 0])
-    angular_vel = np.array([0, 1, 0])
+    angular_vel = np.array([0, 1, 0], dtype=float)
 
-    expected_sol = np.zeros([3, 3])
+    expected_sol = np.zeros([3, 3], dtype=float)
     expected_sol[0, 1] = - 1
     expected_sol[1, 0] = 1
 
@@ -179,7 +179,7 @@ def test2_jac_kinematic_angular_eqs():
 def test1_navigation_eqs():
 
     time = 0
-    linear_vel = np.array([1, 1, 1])
+    linear_vel = np.array([1, 1, 1], dtype=float)
     euler_angles = np.array([np.pi / 4, np.pi / 4, 0])
 
     expected_sol = np.array([1 + (2 ** 0.5) / 2, 0, 1 - (2 ** 0.5) / 2])
@@ -191,10 +191,10 @@ def test1_navigation_eqs():
 def test2_navigation_eqs():
 
     time = 0
-    linear_vel = np.array([1, 0, 1])
+    linear_vel = np.array([1, 0, 1], dtype=float)
     euler_angles = np.array([0, np.pi / 2, 0])
 
-    expected_sol = np.array([1, - 1, 0])
+    expected_sol = np.array([1, - 1, 0], dtype=float)
     sol = navigation_eqs(time, linear_vel, euler_angles)
 
     assert(np.allclose(expected_sol, sol))
